cached angle api calls
