+++ src/apps/angle/ethereum/angle.vaults.contract-position-fetcher.ts
@@ -2,6 +2,7 @@ import { Inject } from '@nestjs/common';
 
 import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import { Register } from '~app-toolkit/decorators';
+import { getImagesFromToken, getLabelFromToken } from '~app-toolkit/helpers/presentation/image.present';
 import { ContractType } from '~position/contract.interface';
 import { PositionFetcher } from '~position/position-fetcher.interface';
 import { ContractPosition } from '~position/position.interface';
@@ -10,46 +11,25 @@ import { Network } from '~types/network.interface';
 
 import { ANGLE_DEFINITION } from '../angle.definition';
 import { AngleContractFactory } from '../contracts';
-import { callAngleApi } from '../helpers/angle.api';
+import { AngleApiHelper } from '../helpers/angle.api';
 
 const appId = ANGLE_DEFINITION.id;
 const groupId = ANGLE_DEFINITION.groups.vaults.id;
 const network = Network.ETHEREUM_MAINNET;
-
-type TVaultManager = {
-  address: string;
-  borrowFee: number;
-  collateral: string;
-  collateralHasPermit: boolean;
-  collateralPermitVersion: string;
-  decimals: number;
-  dust: number;
-  liquidationPenalty: number;
-  maxLTV: number;
-  minCollateralRatio: number;
-  rate: number;
-  stabilityFee: number;
-  stablecoin: string;
-  swapper: string;
-  symbol: string;
-  totalCollateral: number;
-  totalDebt: number;
-  treasury: string;
-};
 
 @Register.ContractPositionFetcher({ appId, groupId, network })
 export class EthereumAngleVaultsContractPositionFetcher implements PositionFetcher<ContractPosition> {
   constructor(
     @Inject(APP_TOOLKIT) private readonly appToolkit: IAppToolkit,
     @Inject(AngleContractFactory) private readonly angleContractFactory: AngleContractFactory,
+    @Inject(AngleApiHelper)
+    private readonly angleApiHelper: AngleApiHelper,
   ) {}
 
   async getPositions() {
     const baseTokenDependencies = await this.appToolkit.getBaseTokenPrices(network);
 
-    const vaultManagers = Object.values(
-      await callAngleApi<Record<string, TVaultManager>>('vaultManagers', { chainId: 1 }),
-    );
+    const vaultManagers = Object.values(await this.angleApiHelper.getVaultManagers());
 
     const positions = vaultManagers.map(vaultManager => {
       const collateralToken = baseTokenDependencies.find(
@@ -74,9 +54,8 @@ export class EthereumAngleVaultsContractPositionFetcher implements PositionFetch
           maxLTV: vaultManager.maxLTV,
         },
         displayProps: {
-          label: `Angle Vault ${vaultManager.symbol}`,
-          images: [''],
-          appName: 'Angle',
+          label: `${getLabelFromToken(collateralToken)}/${getLabelFromToken(stableToken)}`,
+          images: [...getImagesFromToken(collateralToken), ...getImagesFromToken(stableToken)],
         },
       } as ContractPosition;
     });
